#104: Updated error logging behavior, removed Chalk from schema, and reorganized tests
