+++ lib/certification/index.js
@@ -20,7 +20,7 @@ const {
   CURRENT_DATA_DICTIONARY_VERSION,
   DATA_DICTIONARY_VERSIONS,
   CERTIFICATION_FILES,
-  NOT_OK
+  getErrorHandler
 } = require('../../common');
 
 const { REPLICATION_STRATEGIES } = require('../replication/utils');
@@ -157,28 +157,6 @@ const isClientCredentialsConfig = ({ clientCredentials = {} }) =>
  * @returns true if the config is for bearer tokens, false otherwise
  */
 const isBearerTokenConfig = ({ token }) => !!token && token?.length;
-
-/**
- * Creates context-sensitive error handler function that logs to
- * the console or throws errors depending on whether the caller is using the CLI
- *
- * @param {Boolean} fromCli true if coming from the CLI, false otherwise (default)
- * @returns Error handler function
- */
-const getErrorHandler = (fromCli = false) => {
-  return (message, { terminate = true } = {}) => {
-    if (fromCli) {
-      console.error(`${message}`);
-      if (terminate) {
-        process.exit(NOT_OK);
-      } else {
-        process.exitCode = NOT_OK;
-      }
-    } else {
-      throw new Error(message);
-    }
-  };
-};
 
 /**
  * Runs tests for a bulk config file of the user's choosing.
