+++ lib/schema/index.js
@@ -1,12 +1,10 @@
 'use strict';
 
-//TODO remove chalk
-const chalk = require('chalk');
 const { promises: fs } = require('fs');
 const { generateSchema, generateJsonSchema } = require('./generate');
 const path = require('path');
 const { validatePayload, isValidDataDictionaryVersion, validate, combineErrors, VALIDATION_ERROR_MESSAGES } = require('./validate');
-const { CURRENT_DATA_DICTIONARY_VERSION } = require('../../common');
+const { CURRENT_DATA_DICTIONARY_VERSION, getErrorHandler } = require('../../common');
 const { readDirectory } = require('../restore');
 const { getReferenceMetadata } = require('@reso/reso-certification-etl');
 const { processFiles, createDirectoryIfNotPresent, writeFile, readFile } = require('./utils');
@@ -35,16 +33,21 @@ const schema = async ({
   validate,
   version = CURRENT_DATA_DICTIONARY_VERSION,
   payloadPath,
-  resourceName
+  resourceName,
+  fromCli = false
 }) => {
+
+  const handleError = getErrorHandler(fromCli);
+
   try {
-    if ((!generate && !validate) || (generate && validate)) {
-      console.log(chalk.redBright('Only one of --generate (-G) or --validate (-V) should be passed'));
-      return;
+    if (fromCli) {
+      if ((!generate && !validate) || (generate && validate)) {
+        handleError('Only one of --generate (-G) or --validate (-V) should be passed');
+      }
     }
 
     if (!isValidDataDictionaryVersion(version)) {
-      throw new Error(`Invalid Data Dictionary version: '${version}'`);
+      handleError(`Invalid Data Dictionary version: '${version}'`);
     }
 
     if (generate) {
@@ -52,61 +55,66 @@ const schema = async ({
       if (!metadataPath) {
         metadataReport = getReferenceMetadata(version);
         if (!metadataReport) {
-          console.log(chalk.redBright(`Invalid version ${version}`));
-          return;
+          handleError(`Invalid version ${version}`);
         }
       }
 
-      const result = await generateSchemaFromMetadata({
-        metadataPath,
-        metadataReport,
-        additionalProperties
-      });
+      try {
+        const result = await generateSchemaFromMetadata({
+          metadataPath,
+          metadataReport,
+          additionalProperties
+        });
 
-      if (result?.schema) {
-        const { schema } = result;
+        if (result?.schema) {
+          const { schema } = result;
 
-        // write schema to the output path
-        const fileName = 'schema-' + (metadataPath?.split('/')?.at(-1) || 'metadata.json');
-        const success = await writeFile(path.join(outputPath, fileName), JSON.stringify(schema));
-        if (!success) {
-          console.log(chalk.redBright.bold('Error writing generated schema to the given location'));
-          return;
+          // write schema to the output path
+          const fileName = 'schema-' + (metadataPath?.split('/')?.at(-1) || 'metadata.json');
+          const success = await writeFile(path.join(outputPath, fileName), JSON.stringify(schema));
+          if (!success) {
+            handleError('Error writing generated schema to the given location');
+          }
+
+          if (fromCli) {
+            console.log(`Schema successfully generated and saved in ${outputPath}/${fileName}`);
+          }
         }
-        console.log(chalk.greenBright.bold(`Schema successfully generated and saved in ${outputPath}/${fileName}`));
+      } catch (err) {
+        handleError(err);
       }
-
-      return;
-    }
-
-    if (validate) {
+    } else if (validate) {
       if (!payloadPath) {
-        throw new Error('Invalid path to payloads');
+        handleError('Invalid path to payloads');
       }
 
       if (!(resourceName && version)) {
-        throw new Error('Resource name (-r, --resourceName) and version (-v, --version) should be passed together');
+        handleError('Resource name (-r, --resourceName) and version (-v, --version) should be passed together');
       }
 
       await fs.rm(OUTPUT_DIR, { recursive: true, force: true });
 
       const outputPathExists = await createDirectoryIfNotPresent(OUTPUT_DIR);
       if (!outputPathExists) {
-        throw new Error('Unable to create output directory for extracted files');
+        handleError('Unable to create output directory for extracted files');
       }
 
-      await processFiles({ inputPath: payloadPath, outputPath: OUTPUT_DIR });
+      try {
+        await processFiles({ inputPath: payloadPath, outputPath: OUTPUT_DIR });
 
-      await validatePayloads({
-        metadataPath,
-        payloadPath: OUTPUT_DIR,
-        additionalProperties,
-        version,
-        resourceName
-      });
+        await validatePayloads({
+          metadataPath,
+          payloadPath: OUTPUT_DIR,
+          additionalProperties,
+          version,
+          resourceName
+        });
+      } catch (err) {
+        handleError(err);
+      }
     }
   } catch (err) {
-    throw new Error(`Something went wrong! Error: ${err}`);
+    handleError(`Something went wrong! Error: ${err}`);
   }
 };
