+++ lib/variations/index.js
@@ -379,7 +379,7 @@ const computeVariations = async ({
 
           if (!isStandardField) {
             const localFieldNameSuggestions = suggestionsMap?.[resourceName]?.[fieldName]?.suggestions ?? [],
-              hasLocalFieldNameSuggestions = (localFieldNameSuggestions && localFieldNameSuggestions?.length) ?? false;
+              hasLocalFieldNameSuggestions = !!(localFieldNameSuggestions && localFieldNameSuggestions?.length) ?? false;
 
             // try looking in suggestions map
             if (hasLocalFieldNameSuggestions) {
@@ -499,7 +499,7 @@ const computeVariations = async ({
               //if the lookupValue doesn't exist in the standard metadata map then try and find matches
               if (!isStandardLookupValue) {
                 const localLookupValueSuggestions = suggestionsMap?.[resourceName]?.[fieldName]?.[lookupValue]?.suggestions ?? [],
-                  hasLocalLookupValueSuggestions = (localLookupValueSuggestions && localLookupValueSuggestions?.length) ?? false;
+                  hasLocalLookupValueSuggestions = !!(localLookupValueSuggestions && localLookupValueSuggestions?.length) ?? false;
 
                 if (hasLocalLookupValueSuggestions) {
                   // for each suggestion, only use it if the suggested item isn't already there...
